Refactor Dealer test rig to validate mocking stats
